chore: update cli actions file structure
